all routes protected on backend with google auth.
removed body-parser
