+++ api/billing/billingRouter.js
@@ -1,7 +1,7 @@
 require('dotenv');
 const router = require('express').Router();
 // const stripe = require("stripe")(process.env.SK_TEST);
-router.use(require("body-parser").text());
+// router.use(require("body-parser").text());
 const stripe = require("stripe")(process.env.SK_TEST);
 const client = require('twilio')(process.env.ACCOUNT_SID, process.env.AUTH_TOKEN);
 const axios = require('axios');
