+++ api/middleware/auth.js
@@ -1,29 +1,28 @@
-// require('dotenv').config();
+require('dotenv').config();
 
-// let admin = require('firebase-admin');
+let admin = require('firebase-admin');
 
-// let serviceAccount = require(process.env.GOOGLE_APPLICATION_CREDENTIALS);
+let serviceAccount = require(process.env.GOOGLE_APPLICATION_CREDENTIALS);
 
-// admin.initializeApp({
-//     credential: admin.credential.cert(serviceAccount)
-//     // databaseURL: 'https://voicechatroom-1874e.firebaseio.com'
-// });
+admin.initializeApp({
+    credential: admin.credential.cert(serviceAccount)
+});
 
-// const auth = (req, res, next) => {
-//     // var idToken = req.get('Authorization');
-//     var idToken = req.body;
-
-//     // idToken comes from the client app
-//     admin
-//         .auth()
-//         .verifyIdToken(idToken)
-//         .then(function(decodedToken) {
-//             let uid = decodedToken.uid;
-//             res.locals.uid = uid;
-//             next();
-//         })
-//         .catch(function(error) {
-//             res.send(500).json({ message: 'Invalid token' });
-//         });
-// };
-// module.exports = auth;
+function auth(req, res, next) {
+    if (req.headers.application) {
+        admin
+            .auth()
+            .verifyIdToken(req.headers.application)
+            .then(function(decodedToken) {
+                let uid = decodedToken.uid;
+                res.locals.uid = uid;
+                next();
+            })
+            .catch(function(error) {
+                res.status(403).json({ message: 'Invalid token' });
+            });
+    } else {
+        res.status(403).json({ message: 'No authentication token' });
+    }
+}
+module.exports = auth;
