+++ api/server.js
@@ -1,16 +1,8 @@
 require('dotenv').config();
 const express = require('express');
 const cors = require('cors');
-let admin = require('firebase-admin');
-
-let serviceAccount = require(process.env.GOOGLE_APPLICATION_CREDENTIALS);
-
-admin.initializeApp({
-    credential: admin.credential.cert(serviceAccount)
-    // databaseURL: 'https://voicechatroom-1874e.firebaseio.com'
-});
-// var bodyParser = require('body-parser')
-// var jsonParser = bodyParser.json()
+const morgan = require('morgan');
+const auth = require('./middleware/auth');
 
 const teamRouter = require('./team/teamRouter');
 const usersRouter = require('./users/usersRouter');
@@ -22,25 +14,10 @@ const billingRouter = require('./billing/billingRouter');
 const db = require('../data/dbConfig.js');
 const server = express();
 
+server.use(morgan('dev'));
 server.use(cors());
+server.use(auth);
 server.use(express.json());
-// server.use(auth());
-server.use(function(req, res, next) {
-    // var idToken = req.get('Authorization');
-    let { idToken } = req.body;
-    // idToken comes from the client app
-    admin
-        .auth()
-        .verifyIdToken(idToken)
-        .then(function(decodedToken) {
-            let uid = decodedToken.uid;
-            res.locals.uid = uid;
-            next();
-        })
-        .catch(function(error) {
-            res.status(500).json({ message: 'Invalid token' });
-        });
-});
 
 server.use('/api/team', teamRouter);
 server.use('/api/users', usersRouter);
